Fix bugs in godiff pkg and adapt tests according to that, refactor utils function and fix nits in tests
